[#2697] Split prettifyCoordinates and prettifyProjection into separate files.
