+++ lib/ReactViewModels/MouseCoords.js
@@ -6,7 +6,8 @@ import EarthGravityModel1996 from '../Map/EarthGravityModel1996';
 import EllipsoidTerrainProvider from 'terriajs-cesium/Source/Core/EllipsoidTerrainProvider';
 import Intersections2D from 'terriajs-cesium/Source/Core/Intersections2D';
 import knockout from 'terriajs-cesium/Source/ThirdParty/knockout';
-import {prettifyCoordinates, prettifyProjection} from '../Models/prettyCoordinates';
+import prettifyCoordinates from '../Models/prettifyCoordinates';
+import prettifyProjection from '../Models/prettifyProjection';
 import sampleTerrainMostDetailed from 'terriajs-cesium/Source/Core/sampleTerrainMostDetailed';
 import when from 'terriajs-cesium/Source/ThirdParty/when';
